+++ src/app/(sign)/signIn/findPassword/page.tsx
@@ -3,15 +3,9 @@
 import React, { useRef } from 'react'
 import styles from '@/app/(sign)/_components/sign.module.scss'
 import Link from 'next/link'
-import { supabase, executeQuery } from '@/utils/apiRequest/defaultApiSetting'
-
-interface IProfile {
-  email: string
-  id: string
-  password_hint: string
-  profile_img: 0 | 1 | 2 | 3
-  user_name: string
-}
+import { IProfile, getProfile } from '@/utils/apiRequest/profileApiRequest'
+import { supabase } from '@/utils/apiRequest/defaultApiSetting'
+import { BasicInput, PasswordHintInput } from '../../_components/SignInput'
 
 const FindPasswordPage = () => {
   const emailRef = useRef<HTMLInputElement>(null)
@@ -20,28 +14,29 @@ const FindPasswordPage = () => {
   const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault()
 
-    const email = emailRef.current?.value
-    const passwordHint = passwordHintRef.current?.value
+    const email = emailRef.current!.value
+    const passwordHint = passwordHintRef.current!.value
+    const userdata: IProfile[] = await getProfile('email', email)
 
-    const userdata: IProfile[] = await executeQuery(
-      supabase.from('profiles').select('*').eq('email', email),
-      '유저정보를 불러오지 못했습니다',
-    )
     if (!userdata.length) {
       alert('유효하지 않은 이메일 입니다!')
       return null
     }
 
-    const profilePasswordHint = userdata[0].password_hint
-    if (passwordHint !== profilePasswordHint) {
+    if (passwordHint !== userdata[0].password_hint) {
       alert('비밀번호 힌트가 일치하지 않습니다!')
       return null
     }
 
-    await supabase.auth.resetPasswordForEmail(email!, {
-      redirectTo: 'http://localhost:3000/signIn/resetPassword',
-    })
-    alert('비밀번호 재설정 이메일이 발송되었습니다.')
+    try {
+      await supabase.auth.resetPasswordForEmail(email, {
+        redirectTo: 'http://localhost:3000/signIn/resetPassword',
+      })
+      alert('비밀번호 재설정 이메일이 발송되었습니다.')
+    } catch (error) {
+      alert('비밀번호 재설정 시도 중 오류가 발생하였습니다.')
+      throw error
+    }
   }
 
   return (
@@ -49,27 +44,15 @@ const FindPasswordPage = () => {
       <h1>비밀번호 찾기</h1>
 
       <form onSubmit={handleSubmit}>
-        <div className={styles.inputField}>
-          <h3>이메일</h3>
-          <input
-            ref={emailRef}
-            name="email"
-            type="email"
-            placeholder="이메일을 입력해주세요"
-          />
-        </div>
-
-        <div className={styles.inputField}>
-          <h3>비밀번호 찾기 힌트</h3>
-          <input
-            ref={passwordHintRef}
-            name="password_hint"
-            type="text"
-            placeholder="크리스마스에 받았던 가장 좋은 선물은 무엇이었나요?"
-          />
-        </div>
-
-        <button type="submit">완료</button>
+        <BasicInput
+          inputRef={emailRef}
+          title="이메일"
+          name="email"
+          type="email"
+          placeholder="이메일을 입력해주세요."
+        />
+        <PasswordHintInput passwordHintRef={passwordHintRef} />
+        <button type="submit">비밀번호 재설정 이메일 받기</button>
       </form>
 
       <div className={styles.routerBox}>
