+++ src/app/(sign)/signIn/resetPassword/page.tsx
@@ -4,6 +4,7 @@ import React, { useRef } from 'react'
 import styles from '@/app/(sign)/_components/sign.module.scss'
 import { useRouter } from 'next/navigation'
 import { supabase } from '@/utils/apiRequest/defaultApiSetting'
+import { PasswordInput, PasswordCheckInput } from '../../_components/SignInput'
 
 const ResetPasswordPage = () => {
   const router = useRouter()
@@ -25,47 +26,10 @@ const ResetPasswordPage = () => {
   return (
     <main className={styles.container}>
       <h1>비밀번호 재설정</h1>
-
       <form onSubmit={handleSubmit}>
-        <div className={styles.inputField}>
-          <h3>새 비밀번호</h3>
-          <input
-            ref={passwordRef}
-            placeholder="영어,숫자,특수문자 포함 8~15 글자를 입력해주세요"
-            minLength={8}
-            maxLength={15}
-            pattern="(?=.*\d)(?=.*[a-zA-Z])(?=.*[!@#\$%\^*+=\-]).*"
-            required
-            onInvalid={e => {
-              const inputElement = e.currentTarget as HTMLInputElement
-              if (inputElement.validity.patternMismatch)
-                inputElement.setCustomValidity(
-                  '영어,숫자,특수문자 포함 8~15 글자로 입력해주세요',
-                )
-            }}
-            onInput={e => e.currentTarget.setCustomValidity('')}
-            type="password"
-          />
-          <p>특수문자는 다음 문자들만 입력 가능합니다 : !@#$%^*+=- </p>
-        </div>
-
-        <div className={styles.inputField}>
-          <h3>비밀번호 확인</h3>
-          <input
-            type="password"
-            placeholder="비밀번호를 힌번 더 입력해주세요"
-            required
-            onChange={e =>
-              e.currentTarget.setCustomValidity(
-                passwordRef.current?.value !== e.currentTarget.value
-                  ? '비밀번호 확인이 일치하지 않습니다'
-                  : '',
-              )
-            }
-          />
-        </div>
-
-        <button type="submit">비밀번호 변경</button>
+        <PasswordInput passwordRef={passwordRef} />
+        <PasswordCheckInput passwordRef={passwordRef} />
+        <button type="submit">비밀번호 변경 완료</button>
       </form>
     </main>
   )
